Refactored main web application for better modularity and readability
The core change in this commit affects the organization and structure of the main web application file. The previous version contained all the functional components directly in it and was loaded with multiple import statements which were reduced. In the current version, import statements are diminished and the code is neatly modularized to enhance its readability and maintainability. This decision is in accordance with practices promoting clean code and will likely facilitate easier future updates and debugging. Key functional modules such as CSS setup, page configurations and the initialization of the session state have been moved to their respective modules under `src.env_web.modules`. This brings more order to the project structure and eliminates the unnecessary overhead from the main application. There is no mention of affecting any functionalities, so it is understood that this is a refactoring effort maintaining the same functionality while improving the code infrastructure.